refactor(stream): remove old non-append-only Top-N executor
