+++ src/stream/src/from_proto/mod.rs
@@ -33,8 +33,8 @@ mod project;
 mod project_set;
 mod sink;
 mod source;
+mod top_n;
 mod top_n_appendonly;
-mod top_n_new;
 mod union;
 
 // import for submodules
@@ -63,8 +63,8 @@ use self::project::*;
 use self::project_set::*;
 use self::sink::*;
 use self::source::*;
+use self::top_n::*;
 use self::top_n_appendonly::*;
-use self::top_n_new::*;
 use self::union::*;
 use crate::executor::{BoxedExecutor, Executor, ExecutorInfo};
 use crate::task::{ExecutorParams, LocalStreamManagerCore};
